Deprecate legacy plugin class
